Now raw number of guys and girls and implemented a list view
It used to store the percent of guys at a bar, but now it stores the
number of guys and the number of girls.
Fixed map problem by dividing meters by 1000 to get kilometers (for
Parse geopoint call).
Created List view for what user has on map - lists out all the bars,
the number of guys and girls, and the average age for each bar.